+++ modules/apps/frontend-js/frontend-js-spa-web/src/main/resources/META-INF/resources/util/utils.js
@@ -125,6 +125,26 @@ export function getNodeOffset(node) {
 	};
 }
 
+/**
+ * Given a portletId, returns the ID of the portlet's boundary DOM element
+ * @param  {!String} portletId The portlet ID
+ * @return {!String} The portlet boundary ID
+ */
+export function getPortletBoundaryId(portletId) {
+	return 'p_p_id_' + portletId + '_';
+}
+
+/**
+ * Given an array of portlet IDs, returns an array of portlet boundary IDs
+ * @param  {!Array} The collection of portletIds
+ * @return {!Array} The collection of portlet boundary IDs
+ */
+export function getPortletBoundaryIds(portletIds) {
+	return portletIds.map((portletId) => {
+		return getPortletBoundaryId(portletId);
+	});
+}
+
 export function getUid() {
 	return uniqueIdCounter++;
 }
@@ -213,6 +233,27 @@ export function removePathTrailingSlash(path) {
 	}
 
 	return path;
+}
+
+/**
+ * Destroys all rendered portlets on the page
+ */
+export function resetAllPortlets() {
+	getPortletBoundaryIds(Liferay.Portlet.list).forEach((value) => {
+		const portlet = document.querySelector('#' + value);
+
+		if (portlet) {
+			Liferay.Portlet.destroy(portlet);
+
+			portlet.portletProcessed = false;
+		}
+	});
+
+	Liferay.Portlet.readyCounter = 0;
+
+	Liferay.destroyComponents((component, componentConfig) => {
+		return componentConfig.destroyOnNavigate;
+	});
 }
 
 /**
